requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.1.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.12.0
distro==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.1
filelock==3.15.4
frozendict==2.4.4
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
mysql-connector-python==8.4.0
numpy==1.26.4
openai==1.27.0
pandas==2.2.2
peewee==3.17.5
pillow==10.4.0
platformdirs==4.3.8
protobuf==6.31.1
pyarrow==17.0.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
tabulate==0.9.0
together==1.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
websockets==15.0.1
yarl==1.9.4
yfinance==0.2.65
